Revert "Naatil-Evideya-V2"
