+++ package.json
@@ -10,18 +10,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@react-three/drei": "^9.111.3",
-    "@react-three/fiber": "^8.17.5",
-    "bootstrap": "^5.3.3",
-    "clsx": "^2.1.1",
-    "framer-motion": "^11.3.30",
     "lottie-react": "^2.4.0",
     "react": "^18.3.1",
-    "react-bootstrap": "^2.10.4",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.0",
-    "three": "^0.167.1",
-    "three-globe": "^2.31.1"
+    "react-router-dom": "^6.26.0"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
