+++ src/App.css
@@ -2,6 +2,5 @@
     
     margin: 0;
     padding: 0;
-    /* height: 100%; */
     
 }
