--- src/Home/home.css
@@ -1,131 +1,0 @@
-.left{
-    background-image: linear-gradient(to bottom, #FFB200, #FFFFFD);
-    height: 650px;
-  }
-
-  body {
-    overflow-x: hidden;
-    margin-right: 0;
-    margin-bottom: 0;
-  }
-
-/* Small devices (landscape phones, less than 576px) */
-@media (max-width: 575.98px) {
-  .hero {
-   
-    font-size: .5rem;
-  }
-
-  .left {
-    
-    padding: 35px !important;
-    margin-top: 35 !important;
-    margin-bottom: 100px !important;
-  }
-  .left h1{
-    font-size: 30px !important;
-  }
-  .left p{
-    font-size: 16px !important;
-    font-weight: 400 !important;
-  }
-
-  .left h1, p{
-    text-align: start;
-    margin-left: 0 !important;
-  }
-  .left p{
-    margin-bottom: 15px;
-  }
-  .right p,h3 {
-      text-align: start;
-      margin-bottom: 15px;
-    }
-    .right h3{
-      font-weight: 400;
-    }
-
-  .hero-text p {
-    font-size: 1rem;
-    margin-bottom: 0;
-  }
-
-  .city-icon{
-    padding: 25px;
-  }
-
-  .form-container {
-    width: 100%;
-    padding: 20px;
-    margin-bottom: 0;
-  }
-
-  .col-md-6 {
-    margin-bottom: 0 !important; 
-    padding: 1.5rem !important;
-  }
-
-  .cities{
-    margin-top: 0 !important;
-  }
-
-  .features {
-    margin-top: 0 !important;
-    padding-top: 0 !important;
-  }
-  .features p{
-    margin-bottom: 0;
-  }
-
-  .features col{
-    margin-top: .5rem !important;
-    margin-bottom: .5rem !important;
-  }
-
-  .features h1 {
-    font-size: 1.5rem;
-    margin: 0px !important;
-    padding: 1.5rem !important;
-  }
-
-  .features img {
-    max-width: 80px;
-    margin-bottom: 20px;
-  }
-
-  .cities-h2 {
-    margin-bottom: 15px !important;
-    font-size: 30px !important;
-  }
-
-  .features h4 {
-    font-size: 1.2rem;
-  }
-
-  .features p {
-    font-size: 0.9rem;
-    text-align: center;
-  }
-
-  .col {
-    width: 100%;
-    margin-bottom: 20px;
-  }
-
-  .footer {
-    font-size: 0.9rem;
-    margin-top: 0px !important;
-   }
-
-   .footer p{
-    text-align: center !important;
-   }
-  .aboutleft, .aboutright {
-    flex-basis: 100%;
-    
-  }
-  .aboutright p {
-    margin: 10px;
-    text-align: center; 
-  }
-}
