--- src/components/Cities/Cities.css
@@ -1,7 +1,0 @@
-.card-img {
-    width: 100%;
-    height: 202px;
-    object-fit: cover;
-    object-position: top;
-  }
-
