--- src/components/Footer/Footer.jsx
@@ -1,11 +1,0 @@
-// import React from 'react';
-
-const Footer = () => {
-  return (
-    <div className="footer text-center pb-1 pt-1" style={{ color: 'white', backgroundColor: '#000842' }}>
-      <p>© 2024 Naatil-Evideya. All Rights Reserved.</p>
-    </div>
-  );
-};
-
-export default Footer;
