--- src/test/homecss copuy
@@ -1,144 +1,0 @@
-.left{
-    background-image: linear-gradient(to bottom, #FFB200, #FFFFFD);
-    height: 650px;
-  }
-
-.form{
-  border-bottom: 2px solid #828282;
-}
-
-.btn{
-  color: white;
-  font-size: 17px;
-  background-color: #000842;
-  border-radius: 32px;
-}
-.underline-input {
-  border: none;
-  border-bottom: 1px solid #828282;
-  border-radius: 0;
-  box-shadow: none;
-  padding-left: 0;
-  padding-right: 0;
-}
-.features{
-  margin-top: 80px;
-}
-
-.footer {
-  margin-bottom: 0;
-  margin-right: 0; 
-  padding: 10px; 
-}
-
-body {
-  overflow-x: hidden;
-  margin-right: 0;
-}
-
-
-/* Media Queries */
-
-/* Small devices (landscape phones, less than 576px) */
-@media (max-width: 575.98px) {
-  .hero {
-   
-    font-size: .5rem;
-  }
-
-  .left {
-    order: 1;
-    padding: 35px !important;
-    margin-top: 35 !important;
-    margin-bottom: 100px;
-  }
-  .left h1, p{
-    text-align: start;
-    margin-left: 0 !important;
-  }
-  .left p{
-    margin-bottom: 15px;
-  }
-
-  .right {
-    order: 2;
-    }
-    .right p,h3 {
-      text-align: start;
-      margin-bottom: 15px;
-    }
-    .right h3{
-      font-weight: 400;
-    }
-
-  .hero-text p {
-    font-size: 1rem;
-    margin-bottom: 0;
-  }
-
-  .form-container {
-    width: 100%;
-    padding: 20px;
-    margin-bottom: 0;
-  }
-
-  .col-md-6 {
-    margin-bottom: 0 !important; 
-    padding: 1.5rem !important;
-  }
-
-  .features {
-    margin-top: 0 !important;
-    padding-top: 0 !important;
-  }
-  .features p{
-    margin-bottom: 0;
-  }
-
-  .features col{
-    margin-top: .5rem !important;
-    margin-bottom: .5rem !important;
-  }
-
-  .features h1 {
-    font-size: 1.5rem;
-    margin: 0px !important;
-    padding: 1.5rem !important;
-  }
-
-  .features img {
-    max-width: 80px;
-    margin-bottom: 20px;
-  }
-
-  .features h4 {
-    font-size: 1.2rem;
-  }
-
-  .features p {
-    font-size: 0.9rem;
-    text-align: center;
-  }
-
-  .col {
-    width: 100%;
-    margin-bottom: 20px;
-  }
-
-  .footer {
-    font-size: 0.9rem;
-   }
-
-   .footer p{
-    text-align: center !important;
-   }
-  .aboutleft, .aboutright {
-    flex-basis: 100%;
-    
-  }
-  .aboutright p {
-    margin: 10px;
-    text-align: center; 
-  }
-}
-
